Moves main methods into src folder, add credential validation step
